pyproject.toml
[project]
name = "zestify"
version = "1.0.0"
description = "Helping hand with your cooking"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bleach>=6.2.0",
    "fastapi[standard]>=0.116.0",
    "google-genai>=1.25.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "supabase>=2.17.0",
    "types-bleach>=6.2.0.20250514",
    "types-passlib>=1.7.7.20250602",
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "ruff>=0.12.2",
]
